+++ wheel/python/clvm_rs/clvm_tree.py
@@ -1,5 +1,5 @@
 from .clvm_storage import CLVMStorage
-from .deser import deserialize_as_tuples
+from .de import deserialize_as_tuples
 
 from typing import List, Optional, Tuple
